+++ raytracing/ft_get_norm.c
@@ -1,6 +1,6 @@
 #include "../minirt.h"
 
-int ft_get_norm(t_float3 *i)
+float ft_get_norm(t_float3 *i)
 {
 	return ( sqrt(i->x * i->x + i->y * i->y + i->z * i->z) );
 }
